adjust tests and add doc
