+++ tests/test_load_arrow.py
@@ -1,46 +1,22 @@
-
-import pytest
 import inspect
 import io
 import pyarrow as pa
 
-
-from pyonda.load_arrow import load_table_from_arrow_file_buffer, load_table_from_arrow_file
-from pyonda.utils.processing import (
-    arrow_to_processed_pandas, 
-    convert_julia_uuid,
-    check_if_schema_field_has_unsupported_binary_data
+from pyonda.load_arrow import (
+    load_table_from_arrow_file_buffer, 
+    load_table_from_arrow_file, 
+    load_table_from_arrow_file_in_s3
 )
 
 from tests.fixtures import (
-    signal_arrow_table_path, 
+    aws_credentials,
+    signal_arrow_table_path,
+    lpcm_file_path,
     lpcm_zst_file_path,
+    s3,
+    signal_arrow_table_s3_url,
     assert_signal_arrow_dataframes_equal
 )
-
-
-def test_convert_julia_uuid():
-    assert convert_julia_uuid(None) is None, "None input should return a None output"
-    # TODO how to test this
-
-    
-def test_check_if_schema_field_has_unsupported_binary_data():
-    field = pa.field('weird_binary', pa.binary(16))
-    with pytest.raises(ValueError):
-        check_if_schema_field_has_unsupported_binary_data(field)
-
-    field = pa.field('weird_binary', pa.binary(16), metadata={b'ARROW:extension:name': b'JuliaLang.weird_UUID'})
-    with pytest.raises(ValueError):
-        check_if_schema_field_has_unsupported_binary_data(field)
-
-    field = pa.field('weird_binary', pa.binary(16), metadata={b'ARROW:extension:name': b'JuliaLang.UUID'})
-    check_if_schema_field_has_unsupported_binary_data(field)
-
-
-def test_arrow_to_processed_pandas(signal_arrow_table_path):
-    table = pa.ipc.open_file(pa.memory_map(str(signal_arrow_table_path), 'r')).read_all()
-    df_processed = arrow_to_processed_pandas(table)
-    assert_signal_arrow_dataframes_equal(df_processed)
 
 
 def test_load_table_from_arrow_file_default_is_processed():
@@ -74,4 +50,21 @@ def test_load_table_from_arrow_file(signal_arrow_table_path):
 
 def test_load_table_from_arrow_file_as_processed_dataframe(signal_arrow_table_path):
     df = load_table_from_arrow_file(signal_arrow_table_path, processed_pandas=True)
+    assert_signal_arrow_dataframes_equal(df)
+
+
+def test_load_table_from_arrow_file_in_s3_default_is_processed():
+    signature = inspect.signature(load_table_from_arrow_file_in_s3)
+    defaults = {k: v.default for k, v in signature.parameters.items() if v.default is not inspect.Parameter.empty}
+    assert defaults['processed_pandas']==True
+
+
+def test_load_table_from_arrow_file_in_s3(s3, signal_arrow_table_s3_url, signal_arrow_table_path):
+    table = load_table_from_arrow_file_in_s3(signal_arrow_table_s3_url, processed_pandas=False)
+    reference_table = pa.ipc.open_file(pa.memory_map(str(signal_arrow_table_path), 'r')).read_all()
+    assert table == reference_table, "Loaded arrow table is not as expected"
+
+
+def test_load_table_from_arrow_file_in_s3_as_processed_dataframe(s3, signal_arrow_table_s3_url):
+    df = load_table_from_arrow_file_in_s3(signal_arrow_table_s3_url, processed_pandas=True)
     assert_signal_arrow_dataframes_equal(df)
